+++ swarm/libobjc/config/i386/mframe-sysv4.2MP.h
@@ -33,7 +33,7 @@
     { \
       (TYPE)++; \
     } \
-  while (isdigit(*(TYPE))) \
+  while (isdigit((int) *(TYPE))) \
     { \
       (TYPE)++; \
     } \
